# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gram Sabha Management System API",
    "description": "\n## Comprehensive API for Digital Village Governance\n\nThis API serves both Admin Portal (panchayat management) and Citizen Portal (services & issue reporting).\n\n### Features:\n- **Biometric Authentication** - Facial recognition login\n- **Panchayat & Ward Management** - Complete CRUD operations\n- **Issue Tracking** - Rich content reporting with images/audio\n- **Multi-language Support** - English & Hindi\n- **Member Management** - CSV import, bulk operations\n\n### Authentication:\nMost endpoints require JWT Bearer token authentication.\n\n### File Uploads:\nEndpoints support multipart/form-data for images, audio, and documents.\n    ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://egramsabha.empowerpanchayat.org/",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token"
      }
    },
    "schemas": {
...
      "Issue": {
        "title": "Issue",
        "required": [
          "category",
          "subcategory",
          "createdForId",
          "panchayatId",
          "creatorId"
        ],
        "properties": {
          "text": {
            "type": "string",
            "required": false
          },
          "category": {
            "type": "string",
            "enum": [
              "CULTURE_AND_NATURE",
              "INFRASTRUCTURE",
              "EARNING_OPPORTUNITIES",
              "BASIC_AMENITIES",
              "SOCIAL_WELFARE_SCHEMES",
              "OTHER"
            ]
          },
          "subcategory": {
            "type": "string",
            "enum": [
              "FESTIVALS",
              "TREES_AND_FORESTS",
              "SOIL",
              "NATURAL_WATER_RESOURCES",
              "RELIGIOUS_PLACES",
              "LAND",
              "WATER",
              "ENERGY",
              "TRANSPORTATION",
              "COMMUNICATION",
              "AGRICULTURE",
              "ANIMAL_HUSBANDRY",
              "FISHERIES",
              "SMALL_SCALE_INDUSTRIES",
              "MINOR_FOREST_PRODUCE",
              "KHADI_AND_VILLAGE_INDUSTRIES",
              "HEALTH",
              "EDUCATION",
              "HOUSING_AND_SANITATION",
              "SPORTS_AND_ENTERTAINMENT",
              "FOOD",
              "WEAKER_SECTIONS",
              "HANDICAPPED_WELFARE",
              "FAMILY_WELFARE",
              "WOMEN_AND_CHILD_DEVELOPMENT",
              "POVERTY_ALLEVIATION",
              "OTHER"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "URGENT",
              "NORMAL"
            ]
          },
          "createdForId": {
            "type": "schemaobjectid"
          },
          "status": {
            "type": "string",
            "enum": [
              "REPORTED",
              "PICKED_IN_AGENDA",
              "DISCUSSED_IN_GRAM_SABHA",
              "TRANSFERRED",
              "RESOLVED",
              "NO_ACTION_NEEDED"
            ]
          },
          "toBeResolvedBefore": {
            "type": "string",
            "format": "date-time"
          },
          "remark": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "attachment": {
                  "type": "string"
                },
                "filename": {
                  "type": "string"
                },
                "mimeType": {
                  "type": "string"
                },
                "uploadedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "_id": {
                  "type": "string"
                }
              },
              "required": [
                "attachment"
              ]
            }
          },
          "panchayatId": {
            "type": "schemaobjectid"
          },
          "gramSabhaId": {
            "type": "schemaobjectid"
          },
          "creatorId": {
            "type": "schemaobjectid"
          },
          "transcription": {
            "type": "object",
            "description": {
              "default": {}
            },
            "properties": {
              "requestId": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "PROCESSING",
                  "COMPLETED",
                  "FAILED"
                ]
              },
              "text": {
                "type": "string"
              },
              "originalTranscription": {
                "type": "string"
              },
              "enhancedEnglishTranscription": {
                "type": "string"
              },
              "enhancedHindiTranscription": {
                "type": "string"
              },
              "description": {
                "type": "object",
                "properties": {}
              },
              "language": {
                "type": "string"
              },
              "processingMode": {
                "type": "string"
              },
              "transcriptionProvider": {
                "type": "string"
              },
              "providerInfo": {
                "type": "object",
                "properties": {}
              },
              "llmEnhancementStatus": {
                "type": "object",
                "properties": {}
              },
              "requestedAt": {
                "type": "string",
                "format": "date-time"
              },
              "completedAt": {
                "type": "string",
                "format": "date-time"
              },
              "retryCount": {
                "type": "number"
              },
              "lastError": {
                "type": "string"
              }
            }
          },
          "isSummarized": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_id": {
            "type": "string"
          }
        }
      },
...
      "GramSabha": {
        "title": "GramSabha",
        "required": [
          "title",
          "dateTime",
          "location",
          "scheduledDurationHours",
          "status",
          "panchayatId",
          "scheduledById"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "agenda": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                },
                "description": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                },
                "linkedIssues": {
                  "type": "array",
                  "items": {
                    "type": "schemaobjectid"
                  }
                },
                "createdByType": {
                  "type": "string",
                  "enum": [
                    "USER",
                    "SYSTEM"
                  ]
                },
                "createdByUserId": {
                  "type": "schemaobjectid"
                },
                "_id": {
                  "type": "string"
                }
              },
              "required": [
                "title",
                "description",
                "createdByType",
                "createdByUserId"
              ]
            }
          },
          "scheduledDurationHours": {
            "type": "number"
          },
          "meetingLink": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "CANCELLED",
              "UNSCHEDULED",
              "CONCLUDED",
              "IN_PROGRESS",
              "RESCHEDULED"
            ]
          },
          "minutes": {
            "type": "string"
          },
          "meetingNotes": {
            "type": "string"
          },
          "recordingLink": {
            "type": "string"
          },
          "actualDurationMinutes": {
            "type": "number"
          },
          "transcript": {
            "type": "string"
          },
          "conclusion": {
            "type": "string"
          },
          "jioMeetData": {
            "type": "object",
            "required": false,
            "properties": {}
          },
          "agendaGroupId": {
            "type": "string"
          },
          "agendaSigningStatus": {
            "type": "string"
          },
          "agendaSignatories": {
            "type": "object",
            "required": false,
            "properties": {}
          },
          "signedAgendaPDF": {
            "type": "object",
            "properties": {
              "data": {
                "type": "buffer"
              },
              "contentType": {
                "type": "string"
              }
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "schemaobjectid"
            }
          },
          "attendances": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "checkInTime": {
                  "type": "string",
                  "format": "date-time"
                },
                "verificationMethod": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "PRESENT",
                    "ABSENT",
                    "LATE"
                  ]
                },
                "remarks": {
                  "type": "string"
                },
                "userId": {
                  "type": "schemaobjectid"
                }
              },
              "required": [
                "checkInTime",
                "verificationMethod",
                "status",
                "userId"
              ]
            }
          },
          "guests": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "phoneNumber": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "designation": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "phoneNumber"
              ]
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "attachment": {
                  "type": "string"
                },
                "filename": {
                  "type": "string"
                },
                "mimeType": {
                  "type": "string"
                },
                "uploadedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "uploadedAt"
              ]
            }
          },
          "panchayatId": {
            "type": "schemaobjectid"
          },
          "scheduledById": {
            "type": "schemaobjectid"
          },
          "_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/gram-sabha": {
      "post": {
        "tags": [
          "Gram Sabhas"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GramSabha"
              }
            }
          }
        }
      }
    },
...
    "/api/issues": {
      "post": {
        "tags": [
          "Issues"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Issue"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Issues"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/officials": {
      "get": {
        "tags": [
          "Officials"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Officials"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Official"
              }
            }
          }
        }
      }
    },
...
    "/api/panchayats": {
      "get": {
        "tags": [
          "Panchayats"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Panchayats"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Panchayat"
              }
            }
          }
        }
      }
    },
...
    "/api/users/register-face": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
...
  }
}